fix: 카테고리 페이지에 use client 제거
